refactor: Split lib.rs into several more readable files. Merge migrations into one folder.
